requirements.txt
requests==2.32.5
pytest==8.3.5
allure-pytest==2.13.5
openpyxl==3.1.5
pymysql==1.1.1
responses==0.25.3
